V4 regalloc: optimize fixed interval creation.
Instead of creating ranges for all caller saved registers one at a time,
create a single template for all of them. Then change the template for
a (all) specific registers. So, instead of building the ranges every
time for all registers, do it once, and re-use the array. We currently
treat all registers as caller saved, so the ranges can actually be
shared.

Change-Id: I39111ba88be5c49bbab8fe4f2f8c5ad1c89ef282
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>